requirements.txt
aiosqlite==0.22.0
alembic==1.17.2
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
apprise==1.9.6
asgi-lifespan==2.1.0
asn1crypto==1.5.1
asyncpg==0.31.0
attrs==25.4.0
beartype==0.22.9
boto3==1.42.14
botocore==1.42.14
cachetools==6.2.4
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cloudpickle==3.1.2
colorama==0.4.6
coolname==2.2.0
cryptography==46.0.3
cvxopt==1.3.2
cvxpy==1.7.5
dateparser==1.2.2
docker==7.1.0
ecos==2.0.14
exceptiongroup==1.3.1
fakeredis==2.33.0
fastapi==0.126.0
filelock==3.20.1
fsspec==2025.12.0
graphviz==0.21
greenlet==3.3.0
griffe==1.15.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
humanize==4.15.0
hyperframe==6.1.0
idna==3.11
importlib_metadata==8.7.1
Jinja2==3.1.6
jinja2-humanize-extension==0.4.0
jmespath==1.0.1
joblib==1.5.3
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
lupa==2.6
Mako==1.3.10
Markdown==3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
numpy==2.4.0
oauthlib==3.3.1
osqp==1.0.5
opentelemetry-api==1.39.1
opentelemetry-exporter-prometheus==0.60b1
opentelemetry-instrumentation==0.60b1
opentelemetry-sdk==1.39.1
opentelemetry-semantic-conventions==0.60b1
orjson==3.11.5
packaging==25.0
pandas==2.3.3
pathspec==0.12.1
pendulum==3.1.0
platformdirs==4.5.1
pluggy==1.6.0
prefect==3.6.7
prometheus_client==0.23.1
py-key-value-aio==0.3.0
py-key-value-shared==0.3.0
pycparser==2.23
pydantic==2.12.5
pydantic-extra-types==2.10.6
pydantic-settings==2.12.0
pydantic_core==2.41.5
pydocket==0.16.1
Pygments==2.19.2
PyJWT==2.10.1
pyOpenSSL==25.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-json-logger==4.0.0
python-slugify==8.0.4
pytz==2025.2
PyYAML==6.0.3
readchar==4.2.1
redis==7.1.0
referencing==0.37.0
regex==2025.11.3
requests==2.32.5
requests-oauthlib==2.0.0
rfc3339-validator==0.1.4
rich==14.2.0
rpds-py==0.30.0
ruamel.yaml==0.18.17
ruamel.yaml.clib==0.2.15
s3transfer==0.16.0
scipy==1.17.0
scs==3.2.11
semver==3.0.4
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
snowflake-connector-python==4.1.1
sortedcontainers==2.4.0
SQLAlchemy==2.0.45
starlette==0.50.0
text-unidecode==1.3
toml==0.10.2
tomlkit==0.13.3
typer==0.19.2
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
urllib3==2.6.2
uvicorn==0.40.0
websockets==15.0.1
wrapt==1.17.3
zipp==3.23.0
